Add "Strings" feature
You can now specify a Strings attribute for a font, using a filename as
the value:

  [FontX]
  ...
  Strings = arial10.strings
  ...

bmfont2c will then only output characters needed to print these strings.
A new field is added to the fontStyle_t struct, which tells the offset
of each character in the Bitmaps and Widths. If the character was not
output, the offset is -1.